requirements.txt
Django==1.7.4
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
djangorestframework==3.0.5
gunicorn==19.2.1
psycopg2==2.6
static3==0.5.1
wsgiref==0.1.2
